refactor: merge two types import statements
