+++ packages/vibrant-components/src/lib/HStack/HStackProps.ts
@@ -1,7 +1,6 @@
 import type { ResponsiveValue } from '@vibrant-ui/core';
 import { withVariation } from '@vibrant-ui/core';
-import type { StackProps } from '../Stack';
-import type { Alignment } from '../Stack/StackProps';
+import type { Alignment, StackProps } from '../Stack';
 
 export type HStackProps = Omit<StackProps, 'alignHorizontal' | 'alignVertical' | 'direction'> & {
   alignHorizontal?: ResponsiveValue<Exclude<Alignment, 'stretch'>>;
